pyproject.toml
[project]
name = "anchor-text"
version = "0.1.0"
description = "Transform text using the Literacy Bridge Protocol for phonics education"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Anchor Text Team"}
]
keywords = ["literacy", "phonics", "education", "text-processing", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "litellm>=1.50.0",
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "python-docx>=1.1.0",
    "pdfplumber>=0.11.0",
    "reportlab>=4.2.0",
    "odfpy>=1.4.1",
    "striprtf>=0.0.26",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "tiktoken>=0.7.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
    "pre-commit>=3.7.0",
]

[project.scripts]
anchor-text = "anchor_text.cli:app"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
